moved snow related functions to their own file
