+++ porespy/filters/__init__.py
@@ -57,22 +57,22 @@ from .__funcs__ import fftmorphology
 from .__funcs__ import fill_blind_pores
 from .__funcs__ import find_disconnected_voxels
 from .__funcs__ import find_dt_artifacts
-from .__funcs__ import find_peaks
 from .__funcs__ import flood
 from .__funcs__ import hold_peaks
 from .__funcs__ import local_thickness
 from .__funcs__ import nphase_border
 from .__funcs__ import porosimetry
 from .__funcs__ import prune_branches
-from .__funcs__ import reduce_peaks
 from .__funcs__ import region_size
-from .__funcs__ import snow_partitioning
-from .__funcs__ import snow_partitioning_n
-from .__funcs__ import snow_partitioning_parallel
 from .__funcs__ import trim_disconnected_blobs
 from .__funcs__ import trim_extrema
 from .__funcs__ import trim_floating_solid
 from .__funcs__ import trim_nonpercolating_paths
-from .__funcs__ import trim_nearby_peaks
-from .__funcs__ import trim_saddle_points
 from .__funcs__ import trim_small_clusters
+from ._snows import snow_partitioning
+from ._snows import snow_partitioning_n
+from ._snows import snow_partitioning_parallel
+from ._snows import find_peaks
+from ._snows import reduce_peaks
+from ._snows import trim_nearby_peaks
+from ._snows import trim_saddle_points
